refactor(rc-4): consolidate recovery module layout
Move the 431-line orchestration layer from lib/recovery.ts into a new lib/recovery/hook.ts sibling of the existing storage/constants/types submodules, and convert lib/recovery.ts into a pure barrel re-export (mirroring the RC-2 storage split at lib/storage.ts).

Resolves the recovery.ts vs recovery/ fracture flagged in docs/audits/06-filesystem.md and docs/audits/07-refactoring-plan.md#rc-4. Public API surface is preserved: all previous named exports from ./recovery still resolve unchanged.

No behavior changes. Typecheck + lint + 2088 vitest tests pass against the rebuilt tree.
