+++ lib/recovery/index.ts
@@ -10,3 +10,4 @@
 export * from "./types.js";
 export * from "./constants.js";
 export * from "./storage.js";
+export * from "./hook.js";
